chore: move mention parser to web
