+++ packages/shared/package.json
@@ -25,10 +25,6 @@
       "import": "./dist/src/server/auth/apiKeys.js",
       "require": "./dist/src/server/auth/apiKeys.js"
     },
-    "./src/features/comments/mentionParser": {
-      "import": "./dist/src/features/comments/mentionParser.js",
-      "require": "./dist/src/features/comments/mentionParser.js"
-    },
     "./src/features/rbac": {
       "types": "./dist/src/features/rbac/index.d.ts",
       "import": "./dist/src/features/rbac/index.js",
